Converted TclInteger internal representation to a 64-bit Java long; deprecated TclInteger.get() and replaced with TclInteger.getInt() and TclInteger.getLong() throughout the code base; converted Expression and ExprValue classes to  exclusively do 64-bit long math; implemented 'string is wideinteger'; removed TclWideInteger class

+++ src/main/java/tcl/lang/CallFrame.java
@@ -329,12 +329,12 @@ public class CallFrame {
 		curLevel = (interp.varFrame == null) ? 0 : interp.varFrame.level;
 
 		if ((string.length() > 0) && (string.charAt(0) == '#')) {
-			level = Util.getInt(interp, string.substring(1));
+			level = (int)Util.getInt(interp, string.substring(1));
 			if (level < 0) {
 				throw new TclException(interp, "bad level \"" + string + "\"");
 			}
 		} else if ((string.length() > 0) && Character.isDigit(string.charAt(0))) {
-			level = Util.getInt(interp, string);
+			level = (int)Util.getInt(interp, string);
 			level = curLevel - level;
 		} else {
 			level = curLevel - 1;
